pyproject.toml
[tool.poetry]
name = "bigbytes"
version = "0.8.75-dev"
description = "DevEx poetry configuration"
authors = ["Digitrans, Inc."]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"

[tool.poetry.group.dev.dependencies]
mypy = "*"
black = "*"
flake8 = "*"
pre-commit = "*"
pytest = "*"
pytest-cov = "*"
interrogate = "*"
isort = "*"
commitizen = "*"
ipykernel = "^6.17.1"
pandas-stubs = "*"
flake8-bugbear = "*"
sourcery = "*"

[tool.isort]
profile = "black"

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.ruff.lint]
exclude = [
  "bigbytes/data_preparation/templates",
  "bigbytes/orchestration/db/migrations/versions",
]
ignore = []
select = ["E", "F", "W"]

[tool.ruff.lint.per-file-ignores]
"bigbytes/orchestration/pipeline_scheduler.py" = ["F401", "F403"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
